+++ internal/app/wwctl/container/copy/main.go
@@ -18,6 +18,7 @@ func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 	cdp := &wwapiv1.ContainerCopyParameter{
 		ContainerSource:      args[0],
 		ContainerDestination: args[1],
+		Build:                Build,
 	}
 
 	if !container.DoesSourceExist(cdp.ContainerSource) {
@@ -40,6 +41,13 @@ func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 		err = fmt.Errorf("could not duplicate image: %s", err.Error())
 		wwlog.Error(err.Error())
 		return
+	}
+
+	if cdp.Build {
+		err = container.Build(cdp.ContainerDestination, true)
+		if err != nil {
+			return err
+		}
 	}
 
 	wwlog.Info("Container %s successfully duplicated as %s", cdp.ContainerSource, cdp.ContainerDestination)